feat: :lipstick: Cria componente reaproveitavél CardWrapperInfoAmbientes
